#8955 small typo (reviewers patch)

+++ src/sage/matrix/matrix_mod2_dense.pyx
@@ -1172,7 +1172,7 @@ cdef class Matrix_mod2_dense(matrix_dense.Matrix_dense):   # dense or sparse
 
         INPUT:
 
-        -  ``density`` - Integer; proportion (roughly) to be considered for
+        -  ``density`` - float; proportion (roughly) to be considered for
            changes
         -  ``nonzero`` - Bool (default: ``False``); whether the new entries
            are forced to be non-zero
